printing_client/requirements.txt
python-dotenv
supabase
pywin32
